# `Box` component `0` value fix