… ui/components/app/custom-spending-cap/__snapshots__/custom-spending-cap.test.js.snap
… ui/components/app/ledger-instruction-field/__snapshots__/ledger-instruction-field.test.js.snap
… ui/components/app/modals/eth-sign-modal/__snapshots__/eth-sign-modal.test.js.snap
… ui/components/app/modals/export-private-key-modal/__snapshots__/export-private-key-modal.test.js.snap
… ui/components/app/nft-details/__snapshots__/nft-details.test.js.snap
… ui/components/app/signature-request-original/__snapshots__/signature-request-original.test.js.snap
… ui/components/app/signature-request/__snapshots__/signature-request.component.test.js.snap
… ui/components/component-library/button-link/__snapshots__/button-link.test.js.snap
… ui/components/component-library/button/__snapshots__/button.test.js.snap
… ui/components/component-library/form-text-field/__snapshots__/form-text-field.test.js.snap
… ui/components/component-library/input/__snapshots__/input.test.js.snap
… ui/components/component-library/text-field-search/__snapshots__/text-field-search.test.js.snap
… ui/components/component-library/text-field/__snapshots__/text-field.test.js.snap
… ui/components/institutional/compliance-settings/__snapshots__/compliance-settings.test.js.snap
… ui/components/institutional/confirm-remove-jwt-modal/__snapshots__/confirm-remove-jwt-modal.test.js.snap
… ui/components/multichain/account-list-item/__snapshots__/account-list-item.test.js.snap
… ui/components/multichain/detected-token-banner/__snapshots__/detected-token-banner.test.js.snap
… ui/components/multichain/multichain-import-token-link/__snapshots__/multichain-import-token-link.test.js.snap
… ui/components/multichain/network-list-item/__snapshots__/network-list-item.test.js.snap
… ui/components/ui/box/box.test.tsx
… ui/pages/confirm-signature-request/__snapshots__/index.test.js.snap
… ui/pages/confirmation/templates/__snapshots__/add-ethereum-chain.test.js.snap
… ui/pages/confirmation/templates/__snapshots__/switch-ethereum-chain.test.js.snap
… ui/pages/create-account/import-account/__snapshots__/json.test.tsx.snap
… ui/pages/institutional/confirm-add-institutional-feature/__snapshots__/confirm-add-institutional-feature.test.js.snap
… ui/pages/institutional/connect-custody/__snapshots__/account-list.test.js.snap
… ui/pages/institutional/custody/__snapshots__/custody.test.js.snap
… ui/pages/keychains/__snapshots__/reveal-seed.test.js.snap
… ui/pages/onboarding-flow/metametrics/__snapshots__/metametrics.test.js.snap
… ui/pages/settings/experimental-tab/__snapshots__/experimental-tab.test.js.snap