@georgewrmarshall 2023-06-07
RIP 🪦 on your bones a new typescript box update will be built upon. Moved to https://github.com/MetaMask/metamask-extension/pull/19363